Modify import structure for 0.2
