--- apistar/app.py
@@ -1,3 +1,0 @@
-from apistar.frameworks.wsgi import WSGIApp as App
-
-__all__ = ['App']
