+++ apistar/layouts/minimal/app.py
@@ -1,4 +1,5 @@
-from apistar import App, Include, Route
+from apistar import Include, Route
+from apistar.frameworks.wsgi import WSGIApp as App
 from apistar.handlers import docs_urls, static_urls
 
 
@@ -15,3 +16,7 @@ routes = [
 ]
 
 app = App(routes=routes)
+
+
+if __name__ == '__main__':
+    app.main()
