+++ apistar/layouts/minimal/tests.py
@@ -1,5 +1,5 @@
 from apistar.test import TestClient
-from app import welcome
+from app import app, welcome
 
 
 def test_welcome():
@@ -14,7 +14,7 @@ def test_http_request():
     """
     Testing a view, using the test client.
     """
-    client = TestClient()
+    client = TestClient(app)
     response = client.get('http://localhost/')
     assert response.status_code == 200
     assert response.json() == {'message': 'Welcome to API Star!'}
